+++ restful-service/src/main/java/vn/edu/fpt/sba/dto/request/AlbumRequestDTO.java
@@ -10,6 +10,6 @@ public record AlbumRequestDTO(
         String title,
 
         @NotNull(message = "Artist ID is required")
-        Long artistId) {
+        Integer artistId) {
 
 }
