+++ restful-service/src/main/java/vn/edu/fpt/sba/dto/response/ArtistDetailResponseDTO.java
@@ -3,7 +3,7 @@ package vn.edu.fpt.sba.dto.response;
 import java.util.List;
 
 public record ArtistDetailResponseDTO(
-        Long artistId,
+        Integer artistId,
         String name,
         List<AlbumResponseDTO> albums)
 {
