+++ restful-service/src/main/java/vn/edu/fpt/sba/dto/response/ArtistResponseDTO.java
@@ -1,6 +1,6 @@
 package vn.edu.fpt.sba.dto.response;
 
 public record ArtistResponseDTO(
-        Long artistId,
+        Integer artistId,
         String name) {
 }
